do not throw error during clean if shadow dir is absent and move cleanDir to utils
