+++ utils.go
@@ -3,6 +3,7 @@ package main
 import (
 	"io"
 	"os"
+	"path/filepath"
 )
 
 func copyFile(srcFile string, dstFile string) error {
@@ -26,6 +27,25 @@ func moveFile(srcFile string, dstFile string) error {
 	err := os.Rename(srcFile, dstFile)
 	if err != nil {
 		return err
+	}
+	return nil
+}
+
+func cleanDir(dir string) error {
+	d, err := os.Open(dir)
+	if err != nil {
+		return err
+	}
+	defer d.Close()
+	names, err := d.Readdirnames(-1)
+	if err != nil {
+		return err
+	}
+	for _, name := range names {
+		err = os.RemoveAll(filepath.Join(dir, name))
+		if err != nil {
+			return err
+		}
 	}
 	return nil
 }
